+++ CHANGELOG.rst
@@ -5,7 +5,8 @@ Changelog for cfgrib
 0.9.5.5 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Enable ecCodes implicit MULTI-FIELD support by default, needed for NAM Products by NCEP.
+  See: `#45 <https://github.com/ecmwf/cfgrib/issues/45>`_.
 
 
 0.9.5.4 (2019-01-25)